Began implementing attack animation
Player can left click to attack, and a stab animation ensues. Currently animator checks to see if player left clicks, on top of having something currently equipped. If that is the case, an attack will occur. As of now, the left-clicked variable does not reset upon completion of animation so player can attack only once.

Use it wisely I guess....