Wheelchair enemies now play alert animations when their raycast is hit
There is currently no way to get the wheelchair enemies to revert back to their idles after they are done pointing though. Because the alert animation plays when a flag is triggered from a ray hit, I'm not sure how to set that variable back to false since the ray is permanently "has been hit".